Enable markdown for option descriptions.
Hi Ray,

First of all, fantastic job on this library (and your README.md is top-notch 🥇 💯  ). You helped me in my Slack app project so much. Thank you for your amazing work!

I noticed that in version 0.1.3, your library does not convert option descriptions into markdown. Markdown is now supported to option descriptions (for the Checkboxes element in an Actions block, as far as I know).

On my local development environment, I made this proposed change on line 104 and it seemed to work. However, I didn't run any tests to confirm that there are no breaks or regressions.

Also, it might be cleaner to use a single ternary operator or if-statement for the entire `augmentedProps` declaration. It's up to you.

Anyways, keep up the good work! 👍